Get api-server -> kubelet operations working by creating a cluster role binding
The kubelet asks the api server if things should be allowed according to RBAC.
The user used in this case is the CN of the client certificate, which in this case is `api-server`.
So we need to grant the api server access to the kubelet's resources using RBAC.
After some initial confusion, it seems there is a premade ClusterRole for this.
Kubernetes The Hard Way creates its own, which is technically a subset, but if you
have access to the api server there's nothing you can't do anyway (you could just add your own
RBAC rules) so I don't really care.