api-server: Include service ip on the certificate
This solves the bootstrap issue with coredns which needs to talk to the apiserver
without using its dns name.

We could solve this other ways, eg. by hard-coding the ip in coredns then doing a lookup through it,
or having coredns connect to the node ip directly. This was easier.