cert-manager: Use public DNS servers only
cert-manager attempts to fetch the HTTP01 challenge response itself to verify before
telling letsencrypt to go ahead. This fails when using in-cluster DNS because only
the external ingress points at the challenge responder, and the DNS name being verified
will only point to the external ingress when not using in-cluster DNS.

This is an unavoidable problem with trying to validate certificates for service DNS names
using HTTP01. But using external DNS servers is a reasonable solution.